+++ crates/chainio/clients/avsregistry/src/writer.rs
@@ -375,6 +375,35 @@ impl AvsRegistryChainWriter {
         Ok(*tx.tx_hash())
     }
 
+    /// Force a deregistration of an operator from one or more quorums
+    ///
+    /// # Arguments
+    ///
+    /// * `operator_address` - The address of the operator to be ejected
+    /// * `quorum_numbers` - The quorum numbers to eject the operator from
+    ///
+    /// # Returns
+    ///
+    /// * `TxHash` - The transaction hash of the eject operator transaction
+    pub async fn eject_operator(
+        &self,
+        operator_address: Address,
+        quorum_numbers: Bytes,
+    ) -> Result<TxHash, AvsRegistryError> {
+        info!("ejecting operator from quorum with the AVS's registry coordinator");
+        let provider = get_signer(&self.signer.clone(), &self.provider);
+
+        let contract_registry_coordinator =
+            RegistryCoordinator::new(self.registry_coordinator_addr, provider);
+
+        contract_registry_coordinator.ejectOperator(operator_address, quorum_numbers)
+            .send()
+            .await
+            .map_err(AvsRegistryError::AlloyContractError)
+            .inspect(|tx| info!(tx_hash = ?tx, "successfully ejected operator from quorum with the AVS's registry coordinator"))
+            .map(|tx| *tx.tx_hash())
+    }
+
     /// This function is used to update the account identifier of the AVS's RegistryCoordinator.
     ///
     /// # Arguments
@@ -460,7 +489,9 @@ impl AvsRegistryChainWriter {
 mod tests {
 
     use super::AvsRegistryChainWriter;
+    use crate::test_utils::build_avs_registry_chain_reader;
     use crate::test_utils::create_operator_set;
+    use crate::test_utils::OPERATOR_BLS_KEY;
     use alloy::primitives::{Address, Bytes, FixedBytes, U256};
     use eigen_common::{get_provider, get_signer};
     use eigen_crypto_bls::BlsKeyPair;
@@ -801,6 +832,51 @@ mod tests {
 
     #[tokio::test]
     async fn test_eject_operator() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+        let bls_key = OPERATOR_BLS_KEY.to_string();
+        let register_operator_address = FIRST_ADDRESS;
+        let private_key = FIRST_PRIVATE_KEY.to_string();
+        let quorum_nums = Bytes::from([0]);
+
+        let avs_writer =
+            build_avs_registry_chain_writer(http_endpoint.clone(), private_key.clone()).await;
+
+        test_register_operator(
+            &avs_writer,
+            bls_key,
+            quorum_nums.clone(),
+            http_endpoint.clone(),
+        )
+        .await;
+
+        let avs_reader = build_avs_registry_chain_reader(http_endpoint.clone()).await;
+        let is_registered = avs_reader
+            .is_operator_registered(register_operator_address)
+            .await
+            .unwrap();
+        assert!(is_registered);
+
+        let tx_hash = avs_writer
+            .eject_operator(register_operator_address, quorum_nums)
+            .await
+            .unwrap();
+
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+
+        assert!(tx_status);
+
+        let is_registered = avs_reader
+            .is_operator_registered(register_operator_address)
+            .await
+            .unwrap();
+        assert!(!is_registered);
+    }
+
+    #[tokio::test]
+    async fn test_set_eject_operator() {
         let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
         let private_key = FIRST_PRIVATE_KEY.to_string();
         let new_ejector_address = SECOND_ADDRESS;