+++ CHANGELOG.md
@@ -89,6 +89,18 @@ Those changes in added, changed or breaking changes, should include usage exampl
     let tx_hash = avs_writer.set_ejector(new_ejector_address).await.unwrap();
   ````
 
+* Added new method `eject_operator` in `avsregistry/writer` in [#328](https://github.com/Layr-Labs/eigensdk-rs/pull/328).
+
+  ```rust
+    let register_operator_address = address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266");
+    let quorum_nums = Bytes::from([0]);
+
+    let tx_hash = avs_writer
+        .eject_operator(register_operator_address, quorum_nums)
+        .await
+        .unwrap();
+  ```
+
 * Added new method `is_operator_set_quorum` in `avsregistry/writer` in [#296](https://github.com/Layr-Labs/eigensdk-rs/pull/296).
 
   ```rust